:truck: Move status codes enums to enums.py
